changes to backend needed for part 5

+++ part4-bloglist/controllers/blogs.js
@@ -11,34 +11,13 @@ blogsRouter.get('/', async (request, response) => {
     response.json(blogs)
   })
   
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7)
-  }
-  return null
-}
-
 blogsRouter.post('/', async (request, response) => {
-  const body = request.body
-  const token = getTokenFrom(request)
-  const decodedToken = jwt.verify(token, process.env.SECRET)
-  if (!token || !decodedToken.id) {
+  if (!request.user) {
     return response.status(401).json({error: 'token missing or invalid'})
   }
-  const user = await User.findById(decodedToken.id)
 
-  if (body.likes === undefined){
-    body.likes = 0
-  }
-
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes,
-    user: user._id
-  })
+  const user = request.user
+  const blog = new Blog({ ...request.body, user: user.id})
 
   const savedBlog = await blog.save()
   user.blogs = user.blogs.concat(savedBlog._id)
@@ -48,6 +27,17 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blogToDelete = await Blog.findById(request.params.id)
+  if (!blogToDelete) {
+    return response.status(204).end()
+  }
+
+  if (blogToDelete.user && blogToDelete.user.toString() !== request.user.id) {
+    return response.status(401).json({
+      error: 'only creator can delete the blog'
+    })
+  }
+  
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
